ai chatbot correction
